added employeemobiledto for android client

+++ src/Core/Application/ISD/HR/EmployeeManager/Attendances/AttendanceFunctions.cs
@@ -49,7 +49,7 @@ public class AttendanceFunctions
             else
                 dtTimeOut = tsScheduleTimOut;
 
-            return (double)(tsScheduleTimeIn.Equals(TimeSpan.Zero) || tsScheduleTimOut.Equals(TimeSpan.Zero)
+            return (tsScheduleTimeIn.Equals(TimeSpan.Zero) || tsScheduleTimOut.Equals(TimeSpan.Zero)
                 ? (tsActualTimeOut - tsActualTimeIn).TotalHours
                 : (dtTimeOut - dtTimeIn).TotalHours);
         }
